删除之前添加确认弹框

+++ web/src/views/base-info/role-info/role-info.vue
@@ -238,8 +238,19 @@ export default {
     },
 
     deleteInfo () {
-      RoleInfoApi.deleteRoleInfo(this.selectedInfo).then(() => {
-        this.getRoleInfo();
+      this.$confirm('将执行删除操作, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        RoleInfoApi.deleteRoleInfo(this.selectedInfo).then(() => {
+          this.getRoleInfo();
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
       });
     },
