+++ web/src/views/order-form/form-search.vue
@@ -456,9 +456,20 @@ export default {
     },
 
     deleteOrderForm () {
-      orderFormApi.deleteOrderFormInfo(this.selectedInfo).then(() => {
-        this.getOrderForms()
-      })
+      this.$confirm('将执行删除操作, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        orderFormApi.deleteOrderFormInfo(this.selectedInfo).then(() => {
+          this.getOrderForms();
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
+      });
     },
 
     isSelectedInfoInvalid () {
